fix: clients page — server auth guard + fix search placeholder (#421)
1. Split into server component (auth guard + redirect) + client component
2. Server-side auth prevents infinite spinner on expired session
3. Fix ManageView search placeholder: t('botFilterAll') → t('searchPlaceholder')
